ft-Change signin form buttons alignment-#159428325
This commit properly aligns the signin and signup buttons alignment. The main folder has also been renamed from ‘FULLCONNECTION’ to ‘Client’.

ft-#159428325
