+++ .gitignore
@@ -1,4 +1,4 @@
 UI/stylesheets/*.css.map
-FULLCONNECTION/stylesheets/*.css.map
-FULLCONNECTION/javascripts/*.js.map
+Client/stylesheets/*.css.map
+Client/javascripts/*.js.map
 node_modules
