+++ Client/javascripts/script.js
@@ -1,12 +1,15 @@
-import 'babel-polyfill';
+'use strict';
+
 /* global location */
 /* eslint no-restricted-globals: ["off", "location", "event"] */
 /* eslint no-unused-vars: ["error", { "args": "none" }] */
 /* eslint-disable no-unused-vars */
-document.addEventListener('DOMContentLoaded', () => {
+// const baseUrl = 'https://kampkelly-mydiary-api.herokuapp.com/api/v1';
+var baseUrl = 'http://localhost:3000/api/v1';
+document.addEventListener('DOMContentLoaded', function () {
 	// slider animations
 	document.querySelector('#carousel > ul li:nth-child(1) h4').style.color = '#DFAC2C';
-	$('#carousel > ul li:nth-child(1)').on('click', () => {
+	$('#carousel > ul li:nth-child(1)').on('click', function () {
 		$('.img2-container').fadeOut('fast');
 		$('.img3-container').fadeOut('fast');
 		$('.img1-container').fadeIn('fast');
@@ -14,7 +17,7 @@ document.addEventListener('DOMContentLoaded', () => {
 		document.querySelector('#carousel > ul li:nth-child(2) h4').style.color = '#052F60';
 		document.querySelector('#carousel > ul li:nth-child(3) h4').style.color = '#052F60';
 	});
-	$('#carousel > ul li:nth-child(2)').on('click', () => {
+	$('#carousel > ul li:nth-child(2)').on('click', function () {
 		$('.img1-container').fadeOut('fast');
 		$('.img3-container').fadeOut('fast');
 		$('.img2-container').fadeIn('fast');
@@ -22,7 +25,7 @@ document.addEventListener('DOMContentLoaded', () => {
 		document.querySelector('#carousel > ul li:nth-child(1) h4').style.color = '#052F60';
 		document.querySelector('#carousel > ul li:nth-child(3) h4').style.color = '#052F60';
 	});
-	$('#carousel > ul li:nth-child(3)').on('click', () => {
+	$('#carousel > ul li:nth-child(3)').on('click', function () {
 		$('.img1-container').fadeOut('fast');
 		$('.img2-container').fadeOut('fast');
 		$('.img3-container').fadeIn('fast');
@@ -30,8 +33,8 @@ document.addEventListener('DOMContentLoaded', () => {
 		document.querySelector('#carousel > ul li:nth-child(1) h4').style.color = '#052F60';
 		document.querySelector('#carousel > ul li:nth-child(2) h4').style.color = '#052F60';
 	});
-	setInterval(() => {
-		setTimeout(() => {
+	setInterval(function () {
+		setTimeout(function () {
 			$('.img1-container').fadeOut('fast');
 			$('.img3-container').fadeOut('fast');
 			$('.img2-container').fadeIn('fast');
@@ -39,7 +42,7 @@ document.addEventListener('DOMContentLoaded', () => {
 			document.querySelector('#carousel > ul li:nth-child(1) h4').style.color = '#052F60';
 			document.querySelector('#carousel > ul li:nth-child(3) h4').style.color = '#052F60';
 		}, 3000);
-		setTimeout(() => {
+		setTimeout(function () {
 			$('.img1-container').fadeOut('fast');
 			$('.img2-container').fadeOut('fast');
 			$('.img3-container').fadeIn('fast');
@@ -47,7 +50,7 @@ document.addEventListener('DOMContentLoaded', () => {
 			document.querySelector('#carousel > ul li:nth-child(1) h4').style.color = '#052F60';
 			document.querySelector('#carousel > ul li:nth-child(2) h4').style.color = '#052F60';
 		}, 7000);
-		setTimeout(() => {
+		setTimeout(function () {
 			$('.img2-container').fadeOut('fast');
 			$('.img3-container').fadeOut('fast');
 			$('.img1-container').fadeIn('fast');
@@ -59,23 +62,23 @@ document.addEventListener('DOMContentLoaded', () => {
 	// slider animations
 });
 
-document.getElementById('change_settings').addEventListener('click', () => {
+document.getElementById('change_settings').addEventListener('click', function () {
 	document.querySelectorAll('#profile #settings form input[type="time"]')[0].removeAttribute('disabled');
 	document.querySelectorAll('#profile #settings form button')[0].removeAttribute('disabled');
 	document.querySelectorAll('#profile #settings form input[type="time"]')[0].style.backgroundColor = 'white';
 	document.querySelectorAll('#profile #settings form button')[0].style.backgroundColor = '#052F60';
 	document.querySelectorAll('#profile #settings form button')[0].style.color = 'white';
-	document.querySelectorAll('#profile #settings form button')[0].addEventListener('mouseover', () => {
+	document.querySelectorAll('#profile #settings form button')[0].addEventListener('mouseover', function () {
 		document.querySelectorAll('#profile #settings form button')[0].style.backgroundColor = '#8c8c8c';
 	});
-	document.querySelectorAll('#profile #settings form button')[0].addEventListener('mouseleave', () => {
+	document.querySelectorAll('#profile #settings form button')[0].addEventListener('mouseleave', function () {
 		document.querySelectorAll('#profile #settings form button')[0].style.backgroundColor = '#052F60';
 	});
 });
 
 function checkCheckedValues(checkedValues) {
 	// eslint-disable-next-line
-	for (let i = 0; i < checkedValues.length; i++) {
+	for (var i = 0; i < checkedValues.length; i++) {
 		if (checkedValues[i].value === '') {
 			return false;
 		}
@@ -85,9 +88,9 @@ function checkCheckedValues(checkedValues) {
 
 function validateForm(errorMessage, successMessage, event) {
 	event.preventDefault();
-	const req = document.querySelectorAll('form *[required="true"]');
-	const checkedValues = req;
-	if (checkCheckedValues(checkedValues) === true) { // success
+	var req = document.querySelectorAll('form *[required="true"]');
+	var checkedValues = req;
+	if (checkCheckedValues(checkedValues) === true) {// success
 
 	} else {
 		document.querySelector('.form_error_text').style.display = 'block';
@@ -97,114 +100,118 @@ function validateForm(errorMessage, successMessage, event) {
 }
 
 function submitSignup() {
-	const successMessage = 'You have been signed up!';
-	const errorMessage = 'One or more of the required fields is empty!';
-	const noneEmpty = validateForm(errorMessage, successMessage, event);
-	if (noneEmpty === true) { // validation was successful
-		const fullName = document.querySelector('form input[id="signup_fullname"]').value;
-		const dateOfBirth = document.querySelector('form input[id="date_of_birth"]').value;
-		const email = document.querySelector('form input[type="email"]').value.toLowerCase().replace(/\s+/g, '');
-		const password = document.querySelectorAll('form input[type="password"]')[0].value.toLowerCase();
-		const confirmPassword = document.querySelectorAll('form input[type="password"]')[1].value.toLowerCase();
+	var successMessage = 'You have been signed up!';
+	var errorMessage = 'One or more of the required fields is empty!';
+	var noneEmpty = validateForm(errorMessage, successMessage, event);
+	if (noneEmpty === true) {
+		// validation was successful
+		var fullName = document.querySelector('form input[id="signup_fullname"]').value;
+		var dateOfBirth = document.querySelector('form input[id="date_of_birth"]').value;
+		var email = document.querySelector('form input[type="email"]').value.toLowerCase().replace(/\s+/g, '');
+		var password = document.querySelectorAll('form input[type="password"]')[0].value.toLowerCase();
+		var confirmPassword = document.querySelectorAll('form input[type="password"]')[1].value.toLowerCase();
 		if (password !== confirmPassword) {
 			document.querySelector('.form_error_text').style.display = 'block';
 			document.querySelector('.form_error_text small').textContent = 'Your Passwords do not match';
 		} else {
 			document.querySelector('.form_error_text').style.display = 'none';
-			fetch('https://kampkelly-mydiary-api.herokuapp.com/api/v1/auth/signup', {
+			fetch(baseUrl + '/auth/signup', {
 				method: 'POST',
 				headers: {
-					'Content-Type': 'application/json',
+					'Content-Type': 'application/json'
 				},
 				body: window.JSON.stringify({
-					email, password, confirmPassword, dateOfBirth, fullName,
-				}),
-			})
-				.then(res => res.json())
-				.then((data) => {
-					if (data.status === 'Success') {
-						window.location = `signin.html?notice=${data.message}`;
-					} else {
-						document.querySelector('.form_error_text').style.display = 'block';
-						document.querySelector('.form_error_text small').textContent = data.message;
-					}
-				});
+					email: email, password: password, confirmPassword: confirmPassword, dateOfBirth: dateOfBirth, fullName: fullName
+				})
+			}).then(function (res) {
+				return res.json();
+			}).then(function (data) {
+				if (data.status === 'Success') {
+					window.location = 'signin.html?notice=' + data.message;
+				} else {
+					document.querySelector('.form_error_text').style.display = 'block';
+					document.querySelector('.form_error_text small').textContent = data.message;
+				}
+			});
 		}
 	}
 	validateForm();
 }
 
 function submitSignin() {
-	const successMessage = 'You have successfully signed in!';
-	const errorMessage = 'One or more of the required fields is empty!';
-	const noneEmpty = validateForm(errorMessage, successMessage, event);
-	if (noneEmpty === true) { // validation was successful
-		const email = document.querySelector('form input[type="email"]').value.toLowerCase().replace(/\s+/g, '');
-		const password = document.querySelectorAll('form input[type="password"]')[0].value.toLowerCase();
-		fetch('https://kampkelly-mydiary-api.herokuapp.com/api/v1/auth/login', {
+	var successMessage = 'You have successfully signed in!';
+	var errorMessage = 'One or more of the required fields is empty!';
+	var noneEmpty = validateForm(errorMessage, successMessage, event);
+	if (noneEmpty === true) {
+		// validation was successful
+		var email = document.querySelector('form input[type="email"]').value.toLowerCase().replace(/\s+/g, '');
+		var password = document.querySelectorAll('form input[type="password"]')[0].value.toLowerCase();
+		fetch(baseUrl + '/auth/login', {
 			method: 'POST',
 			headers: {
-				'Content-Type': 'application/json',
+				'Content-Type': 'application/json'
 			},
-			body: window.JSON.stringify({ email, password }),
-		})
-			.then(res => res.json())
-			.then((data) => {
-				document.querySelector('.form_error_text').style.display = 'none';
-				if (data.status === 'Success') {
-					localStorage.setItem('diary_token', data.data.token);
-					window.location = `dashboard.html?notice=${data.message}`;
-				} else {
-					document.querySelector('.form_error_text').style.display = 'block';
-					document.querySelector('.form_error_text small').textContent = data.message;
-				}
-			});
+			body: window.JSON.stringify({ email: email, password: password })
+		}).then(function (res) {
+			return res.json();
+		}).then(function (data) {
+			document.querySelector('.form_error_text').style.display = 'none';
+			if (data.status === 'Success') {
+				localStorage.setItem('diary_token', data.data.token);
+				window.location = 'dashboard.html?notice=' + data.message;
+			} else {
+				document.querySelector('.form_error_text').style.display = 'block';
+				document.querySelector('.form_error_text small').textContent = data.message;
+			}
+		});
 	}
 	validateForm();
 }
 
 function forgotPassword() {
-	const successMessage = 'Your password has been sent to the email if it exists!';
-	const errorMessage = 'Please enter a valid email!';
-	const noneEmpty = validateForm(errorMessage, successMessage, event);
-	if (noneEmpty === true) { // validation was successful
-		const email = document.querySelector('form input[type="email"]').value.toLowerCase().replace(/\s+/g, '');
+	var successMessage = 'Your password has been sent to the email if it exists!';
+	var errorMessage = 'Please enter a valid email!';
+	var noneEmpty = validateForm(errorMessage, successMessage, event);
+	if (noneEmpty === true) {
+		// validation was successful
+		var email = document.querySelector('form input[type="email"]').value.toLowerCase().replace(/\s+/g, '');
 	}
 }
 
 function addEntry() {
-	const successMessage = 'Entry has been saved!';
-	const errorMessage = 'One or more of the required fields is empty!';
-	const noneEmpty = validateForm(errorMessage, successMessage, event);
-	if (noneEmpty === true) { // validation was successful
+	var successMessage = 'Entry has been saved!';
+	var errorMessage = 'One or more of the required fields is empty!';
+	var noneEmpty = validateForm(errorMessage, successMessage, event);
+	if (noneEmpty === true) {// validation was successful
 
 	}
 }
 
 function editEntry() {
-	const successMessage = 'Entry has been updated!';
-	const errorMessage = 'One or more of the required fields is empty!';
-	const noneEmpty = validateForm(errorMessage, successMessage, event);
-	if (noneEmpty === true) { // validation was successful
+	var successMessage = 'Entry has been updated!';
+	var errorMessage = 'One or more of the required fields is empty!';
+	var noneEmpty = validateForm(errorMessage, successMessage, event);
+	if (noneEmpty === true) {// validation was successful
 
 	}
 }
 
 function updateProfile() {
-	const successMessage = 'Profile Updated!';
-	const errorMessage = 'Please fill the form fields!';
-	const noneEmpty = validateForm(errorMessage, successMessage, event);
-	if (noneEmpty === true) { // validation was successful
+	var successMessage = 'Profile Updated!';
+	var errorMessage = 'Please fill the form fields!';
+	var noneEmpty = validateForm(errorMessage, successMessage, event);
+	if (noneEmpty === true) {// validation was successful
 
 	}
 }
 
 function saveSettings() {
-	const successMessage = 'Settings Saved!';
-	const errorMessage = 'Please Enter a Valid Date!';
-	const noneEmpty = validateForm(errorMessage, successMessage, event);
-	if (noneEmpty === true) { // validation was successful
-		const time = document.querySelector('form input[type="time"]').value;
+	var successMessage = 'Settings Saved!';
+	var errorMessage = 'Please Enter a Valid Date!';
+	var noneEmpty = validateForm(errorMessage, successMessage, event);
+	if (noneEmpty === true) {
+		// validation was successful
+		var time = document.querySelector('form input[type="time"]').value;
 		document.querySelector('.success-text #notification_time').textContent = time;
 	}
 }
@@ -225,9 +232,9 @@ function closeNotification() {
 function checkNotice() {
 	console.log(window.location.href);
 	console.log('ddlmd');
-	const pageUrl = window.location.href;
-	const url = new URL(pageUrl);
-	const notice = url.searchParams.get('notice');
+	var pageUrl = window.location.href;
+	var url = new URL(pageUrl);
+	var notice = url.searchParams.get('notice');
 	console.log(notice);
 	if (notice) {
 		if (notice.length >= 1) {
@@ -237,3 +244,4 @@ function checkNotice() {
 	}
 }
 checkNotice();
+//# sourceMappingURL=script.js.map
