+++ Client/signup.html
@@ -100,8 +100,8 @@
                                 </div>
                             </div>
 							<div class="field grid-container">
-	                                <div class="col-1-6 col-6-6-md col-6-6-xs"></div>
-	                                <div class="col-3-6 col-6-6-md col-6-6-xs">
+	                                <div class="col-2-6 col-6-6-md col-6-6-xs"></div>
+	                                <div class="col-2-6 col-6-6-md col-6-6-xs">
 	                                    <button type="submit" class="submit-button button-white" id="dkkd" onclick="submitSignup()">Create Account</button>
 	                                </div>
 									<div class="col-2-6 col-6-6-md col-6-6-xs mt-4-md mt-4-xs">
