+++ Client/stylesheets/forms.scss
@@ -19,7 +19,8 @@ form {
 		margin-bottom: 10px;
 	}
     .field {
-        margin: 0px 0px 15px 0px;
+        margin: 0px auto 15px auto;
+		width: 90%;
         label {
             margin: 3px 0px 3px 0px;
             padding: 6px 6px 6px 0px;
@@ -36,7 +37,7 @@ form {
             border: none;
             float: left;
             height: 2.7em;
-            width: 75%;
+            width: 100%;
             border-radius: 3px;
             padding: 0px 6px 0px 6px;
             border-color: inherit;
