+++ Client/stylesheets/signup.scss
@@ -22,19 +22,31 @@ main {
         width: 61%;
         margin: auto;
     }
+	.submit-button {
+		width: 100%;
+	}
     @media #{$smartphone} {
         .form-cover {
             width: 90%;
             margin: auto;
         }
+		.submit-button {
+			width: 100%;
+		}
     }
     @media #{$tablet} {
         .form-cover {
             width: 80%;
         }
 		button.button-white {
-			width: 79%;
-			margin-left: 20px;
+			// width: 79%;
+			// margin-left: 20px;
+		}
+		.submit-button {
+			width: 100%;
+		}
+		[class*='col-'] {
+		  padding-right: 0;
 		}
     }
 }
